refactor: split e2e_test.rs
