+++ crates/pluginlab/tests/e2e_rust_repl_logic.rs
@@ -1,8 +1,11 @@
+mod utils;
+
 #[cfg(test)]
-mod e2e_test {
+mod e2e_rust_repl_logic {
+
+    use crate::utils::*;
 
     use rexpect::spawn;
-    use std::path::PathBuf;
 
     const TEST_TIMEOUT: u64 = 10000;
 
@@ -21,30 +24,6 @@ mod e2e_test {
         });
         println!("Running command: {}", command);
         command
-    }
-
-    fn find_project_root() -> PathBuf {
-        let mut current = std::env::current_dir().unwrap();
-        println!("Starting search from: {:?}", current);
-
-        // Walk up the directory tree looking for the workspace root Cargo.toml
-        loop {
-            let cargo_toml = current.join("Cargo.toml");
-            if cargo_toml.exists() {
-                // Check if this is the workspace root by looking for [workspace] section
-                if let Ok(content) = std::fs::read_to_string(&cargo_toml) {
-                    if content.contains("[workspace]") {
-                        println!("Found workspace root at: {:?}", current);
-                        return current;
-                    }
-                }
-            }
-
-            if !current.pop() {
-                // current.pop() moves up one directory in the path. If we're already at the root, it returns false.
-                panic!("Could not find workspace root (Cargo.toml with [workspace])");
-            }
-        }
     }
 
     #[test]
@@ -217,92 +196,6 @@ mod e2e_test {
         session
             .exp_string("0")
             .expect("Didn't get expected echo output for $? after previous command succeeded");
-    }
-
-    #[test]
-    fn test_ls_plugin() {
-        let project_root = find_project_root();
-        println!("Setting current directory to: {:?}", project_root);
-        std::env::set_current_dir(&project_root).unwrap();
-        let mut session = spawn(
-            &format!(
-                "{} --dir tmp/filesystem --allow-read",
-                &build_command(&["plugin_ls.wasm"], "repl_logic_guest.wasm")
-            ),
-            Some(TEST_TIMEOUT),
-        )
-        .expect("Can't launch pluginlab with plugin greet");
-
-        session
-            .exp_string("[Host] Starting REPL host...")
-            .expect("Didn't see startup message");
-        session
-            .exp_string("[Host] Loading plugin:")
-            .expect("Didn't see plugin loading message");
-        session
-            .exp_string("repl(0)>")
-            .expect("Didn't see REPL prompt");
-        session.send_line("ls").expect("Failed to send command");
-        session
-            .exp_string(
-                "D\tdata\r\nD\tdocuments\r\nD\tlogs\r\nF\t.config\r\nF\t.hidden_file\r\nF\tREADME.md\r\n",
-            )
-            .expect("Didn't get listing of current directory");
-        session
-            .send_line("ls documents")
-            .expect("Failed to send command");
-        session
-            .exp_string("D\tdocuments/work\r\nF\tdocuments/README.md\r\nF\tdocuments/config.json\r\nF\tdocuments/notes.txt\r\n")
-            .expect("Didn't get listing of documents directory");
-        session
-            .send_line("ls documents/work/projects")
-            .expect("Failed to send command");
-        session
-            .exp_string("D\tdocuments/work/projects/alpha\r\nD\tdocuments/work/projects/beta\r\nF\tdocuments/work/projects/.gitkeep\r\n")
-            .expect("Didn't get listing of projects directory");
-        session
-            .send_line("ls data/sample.csv")
-            .expect("Failed to send command");
-        session
-            .exp_string("F\tdata/sample.csv\r\n")
-            .expect("Didn't get listing for a single file");
-    }
-
-    #[test]
-    fn test_cat_plugin() {
-        let project_root = find_project_root();
-        println!("Setting current directory to: {:?}", project_root);
-        std::env::set_current_dir(&project_root).unwrap();
-        let mut session = spawn(
-            &format!(
-                "{} --dir tmp/filesystem --allow-read",
-                &build_command(&["plugin_cat.wasm"], "repl_logic_guest.wasm")
-            ),
-            Some(TEST_TIMEOUT),
-        )
-        .expect("Can't launch pluginlab with plugin greet");
-
-        session
-            .exp_string("[Host] Starting REPL host...")
-            .expect("Didn't see startup message");
-        session
-            .exp_string("[Host] Loading plugin:")
-            .expect("Didn't see plugin loading message");
-        session
-            .exp_string("repl(0)>")
-            .expect("Didn't see REPL prompt");
-        session
-            .send_line("cat documents")
-            .expect("Failed to send command");
-        session
-            .exp_string("cat: documents: Is a directory")
-            .expect("Didn't get expected error output for trying to cat a directory");
-        session
-            .send_line("cat documents/README.md")
-            .expect("Failed to send command");
-        session
-            .exp_string("# Documents")
-            .expect("Didn't get expected contents of README.md");
     }
 
     #[test]
